Various improvements.
In resource QlikVirtualProxy:
- loadBalancingServerNodes : logs and Noop if selection empty
- proxyFilter : the proxy list gives an explicit list of hostnames,
proxyFilter lists proxies that matches a filter expression (same design
as loadBalancingServerNodes for proxies)

In resource QlikNode:
The engine/proxy ... params are not mandatory and not nullable. So its default
value is false. Plus, it does not seem possible to unactivate a service that
has been once activated (Bad request, status 400).

I enforced the fact that if they are false, it does no update. If they are
true, it activates the service.